+++ tests/src/length_prefix.rs
@@ -21,6 +21,14 @@ pub struct Prefix {
     pub reason_length: u8,
 }
 
+#[derive(Protocol, Debug, PartialEq, Eq)]
+pub struct WithElementsLength {
+    pub count: u32,
+    pub foo: bool,
+    #[protocol(length_prefix(elements(count)))]
+    pub data: Vec<u32>,
+}
+
 #[test]
 fn can_read_length_prefix_5_bytes_string() {
     assert_eq!(Foo {
@@ -37,5 +45,21 @@ fn can_read_length_prefix_8_bytes_u32_array() {
         other: 123,
         reason: vec![0x00ff00ff, 0x00ff00ff],
     }, Foo::from_raw_bytes(&[0, 8, 123, 0, !0, 0, !0, 0, !0, 0, !0], &Settings::default()).unwrap());
+}
+
+
+#[test]
+fn can_read_length_prefix_3_elements() {
+    assert_eq!(WithElementsLength {
+        count: 3,
+        foo: true,
+        data: vec![1, 2, 3],
+    }, WithElementsLength::from_raw_bytes(
+                           &[0, 0, 0, 3, // disjoint length prefix
+                             1, // boolean true
+                             0, 0, 0, 1, // 1
+                             0, 0, 0, 2, // 2
+                             0, 0, 0, 3], // 3
+                             &Settings::default()).unwrap());
 }